python/url-shortner/requirements.txt
aws-cdk-lib==2.164.1
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
aws-cdk.aws-dynamodb